+++ packages/widgetbook/lib/src/fields/color_field/color_picker.dart
@@ -31,7 +31,7 @@ class _ColorPickerState extends State<ColorPicker> {
     super.initState();
     // Color.alpha was deprecated in Flutter 3.27.0, the alternative
     // api (.a) is not available in Color for our minimum
-    // Flutter version (3.16.0), as they were also introduced in 3.27.0.
+    // Flutter version (3.19.0), as they were also introduced in 3.27.0.
     // ignore: deprecated_member_use
     alpha = widget.value.alpha;
     colorSpace = widget.colorSpace;